Css changes, broke fleet click handling.
changed invaders/gun to font in hopes of simplifying code.  Was trying to
get char to change with each step of movement but no luck just yet.  Also
click handler is unreliable and noticed whole fleet re renders when clicking
on one alien.  Still, on the right track with animation technique, so will
merge with master and create new branches for other issues.