Successfully add sound to walking aliens using howlerjs
Browsers won't play sound objects unless they are created in response
to a user interaction with the browser, so I added a start button,
which was going to happen anyway. Than provides a click event and we have
sounds playing.  Sounds need to be in the 'public/assets' folder to work.