Progress implementing shot sound.
Todo - devise a workaround so that the kb events will stop interfering with
each other. Will probably have to keep track of which direction keys are down
and when or some such thing.  Still, in a crude sense the shots are making a
sound. Not a great one, but a sound.